update consensus algos to use types in base/
