--- types/src/lib.rs
@@ -1,4 +1,0 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod protocol_state;
